refactor, move overlays to separate folder
