+++ src/shared/components/VideoPlayer/VideoOverlays/index.ts
@@ -1,0 +1,3 @@
+export * from './EndingOverlay'
+export * from './ErrorOverlay'
+export * from './LoadingOverlay'
